Reduces Serializer/Parser interface members to only thos items that are required by the Reader/Writer.
