+++ src/CsvHelper.Tests/CsvHelper.Tests.csproj
@@ -22,6 +22,10 @@
 		  <Reference Include="Microsoft.CSharp" />
 		</ItemGroup>
 
+		<ItemGroup>
+		  <Service Include="{82a7f48d-3b50-4b1e-b82e-3ada8210c358}" />
+		</ItemGroup>
+
 		<PropertyGroup Condition=" '$(TargetFramework)' == 'netcoreapp1.1' ">
 				<DefineConstants>$(DefineConstants);NETCORE</DefineConstants>
 		</PropertyGroup>
