+++ src/CsvHelper/ICsvParser.cs
@@ -2,26 +2,20 @@
 // This file is a part of CsvHelper and is dual licensed under MS-PL and Apache 2.0.
 // See LICENSE.txt for details or visit http://www.opensource.org/licenses/ms-pl.html for MS-PL and http://opensource.org/licenses/Apache-2.0 for Apache 2.0.
 // http://csvhelper.com
-using System;
+
 using System.IO;
-using CsvHelper.Configuration;
 
 namespace CsvHelper
 {
 	/// <summary>
 	/// Defines methods used the parse a CSV file.
 	/// </summary>
-	public interface ICsvParser : IDisposable
+	public interface ICsvParser : IParser
 	{
 		/// <summary>
 		/// Gets the <see cref="TextReader"/>.
 		/// </summary>
 		TextReader TextReader { get; }
-
-		/// <summary>
-		/// Gets the configuration.
-		/// </summary>
-		ICsvParserConfiguration Configuration { get; }
 
 		/// <summary>
 		/// Gets the character position that the parser is currently on.
@@ -33,10 +27,6 @@ namespace CsvHelper
 		/// </summary>
 		long BytePosition { get; }
 
-		/// <summary>
-		/// Gets the row of the CSV file that the parser is currently on.
-		/// </summary>
-		int Row { get; }
 
 		/// <summary>
 		/// Gets the row of the CSV file that the parser is currently on.
@@ -48,11 +38,5 @@ namespace CsvHelper
 		/// Gets the raw row for the current record that was parsed.
 		/// </summary>
 		string RawRecord { get; }
-
-		/// <summary>
-		/// Reads a record from the CSV file.
-		/// </summary>
-		/// <returns>A <see cref="T:String[]" /> of fields for the record read.</returns>
-		string[] Read();
 	}
 }
