+++ src/CsvHelper/ICsvReaderRow.cs
@@ -3,7 +3,6 @@
 // See LICENSE.txt for details or visit http://www.opensource.org/licenses/ms-pl.html for MS-PL and http://opensource.org/licenses/Apache-2.0 for Apache 2.0.
 // http://csvhelper.com
 using System;
-using System.Collections.Generic;
 using CsvHelper.Configuration;
 using CsvHelper.TypeConversion;
 
@@ -157,8 +156,8 @@ namespace CsvHelper
 		T GetField<T>( string name );
 
 		/// <summary>
-		/// Gets the field converted to <see cref="Type"/> T at position 
-		/// (column) name and the index instance of that field. The index 
+		/// Gets the field converted to <see cref="Type"/> T at position
+		/// (column) name and the index instance of that field. The index
 		/// is used when there are multiple columns with the same header name.
 		/// </summary>
 		/// <typeparam name="T"></typeparam>
@@ -188,8 +187,8 @@ namespace CsvHelper
 		T GetField<T>( string name, ITypeConverter converter );
 
 		/// <summary>
-		/// Gets the field converted to <see cref="Type"/> T at position 
-		/// (column) name and the index instance of that field. The index 
+		/// Gets the field converted to <see cref="Type"/> T at position
+		/// (column) name and the index instance of that field. The index
 		/// is used when there are multiple columns with the same header name.
 		/// </summary>
 		/// <typeparam name="T">The <see cref="Type"/> of the field.</typeparam>
@@ -220,8 +219,8 @@ namespace CsvHelper
 		T GetField<T, TConverter>( string name ) where TConverter : ITypeConverter;
 
 		/// <summary>
-		/// Gets the field converted to <see cref="Type"/> T at position 
-		/// (column) name and the index instance of that field. The index 
+		/// Gets the field converted to <see cref="Type"/> T at position
+		/// (column) name and the index instance of that field. The index
 		/// is used when there are multiple columns with the same header name.
 		/// </summary>
 		/// <typeparam name="T">The <see cref="Type"/> of the field.</typeparam>
@@ -250,8 +249,8 @@ namespace CsvHelper
 		bool TryGetField( Type type, string name, out object field );
 
 		/// <summary>
-		/// Gets the field converted to <see cref="System.Type"/> T at position 
-		/// (column) name and the index instance of that field. The index 
+		/// Gets the field converted to <see cref="System.Type"/> T at position
+		/// (column) name and the index instance of that field. The index
 		/// is used when there are multiple columns with the same header name.
 		/// </summary>
 		/// <param name="type">The <see cref="System.Type"/> of the field.</param>
@@ -314,8 +313,8 @@ namespace CsvHelper
 		bool TryGetField<T>( string name, out T field );
 
 		/// <summary>
-		/// Gets the field converted to <see cref="Type"/> T at position 
-		/// (column) name and the index instance of that field. The index 
+		/// Gets the field converted to <see cref="Type"/> T at position
+		/// (column) name and the index instance of that field. The index
 		/// is used when there are multiple columns with the same header name.
 		/// </summary>
 		/// <typeparam name="T"></typeparam>
