+++ src/CsvHelper/ICsvSerializer.cs
@@ -2,31 +2,19 @@
 // This file is a part of CsvHelper and is dual licensed under MS-PL and Apache 2.0.
 // See LICENSE.txt for details or visit http://www.opensource.org/licenses/ms-pl.html for MS-PL and http://opensource.org/licenses/Apache-2.0 for Apache 2.0.
 // http://csvhelper.com
-using System;
+
 using System.IO;
-using CsvHelper.Configuration;
 
 namespace CsvHelper
 {
 	/// <summary>
 	/// Defines methods used to serialize data into a CSV file.
 	/// </summary>
-	public interface ICsvSerializer : IDisposable
+	public interface ICsvSerializer : ISerializer
 	{
 		/// <summary>
 		/// Gets the <see cref="TextWriter"/>.
 		/// </summary>
 		TextWriter TextWriter { get; }
-
-		/// <summary>
-		/// Gets the configuration.
-		/// </summary>
-		ICsvSerializerConfiguration Configuration { get; }
-
-		/// <summary>
-		/// Writes a record to the CSV file.
-		/// </summary>
-		/// <param name="record">The record to write.</param>
-		void Write( string[] record );
 	}
 }
